Summary: Fixed issue 993.
The problem was that Eclipse accepted partial messages sent from MoRE, so if a message was larger than the write buffer then Eclipse would only read the first chunk, and wait for a new message. While Eclipse was waiting MoRE was still trying to write the first message, causing it to hang because pipe buffer was full. The problem surfaced on OS X because it has a rather small buffer size for pipes.

Fixed by making Eclipse wait for the whole message.