Summary: Fixed issue 986 (error during help parsing)
While parsing the help documentation the MoSyncDocsSearchParticipant threw an exception, due to error in parsing. It turned out that it tried to parse a binary file, so now it only parses the html files.

While fixing the issue I took the oppertunity to refactor the help plugin. The following changes were made:
1. Removed com.mobilesorcery.sdk.docs plugin since it seems more reasonable to keep the help code in a single place. It seems like this was the purpose from the beginning, but due to some problem in an earlier Eclipse version it was not possible.
2. Updated the MoSyncDocsSearchParticipant to extend SearchParticipant instead of the now deprecated LuceneSearchParticipant.